Release v1.1.0.24 - Stacker crash fixes
- Fixed transition checker segfault: segmentation map dimension mismatch
  when image exceeds segmentation resolution (nearest-neighbor upscaling)
- Fixed outlierMask truncation: uint32_t -> uint64_t in PixelSelectionResult
  to match PixelStackMetadata (supports 64-frame outlier tracking)
- Added channel count validation when loading frames
- Exception-safe FITS keyword numeric parsing (catches malformed values)

+++ src/engine/PixelSelector.h
@@ -38,7 +38,7 @@ struct PixelSelectionResult
    float confidence = 0.0f;                   // Confidence in selection (0-1)
    RegionClass regionClass = RegionClass::Background;  // ML segmentation class
    float classConfidence = 0.0f;              // Confidence in ML class
-   uint32_t outlierMask = 0;                  // Which frames were rejected
+   uint64_t outlierMask = 0;                  // Which frames were rejected (up to 64)
 
    // Distribution info (for debugging/analysis)
    float distMu = 0.0f;
